tcp service. Leader health check

+++ Distributed-Data-Store/Server1/Services/DistributionService/DistributionService.cs
@@ -11,25 +11,26 @@ public class DistributionService : IDistributionService
 {
     private readonly IDataService _dataService;
     private readonly IHttpService _httpService;
+    private readonly ITcpService _tcpService;
 
-    public DistributionService(IDataService dataService, IHttpService httpService)
+    public DistributionService(IDataService dataService, IHttpService httpService, ITcpService tcpService)
     {
         _dataService = dataService;
         _httpService = httpService;
+        _tcpService = tcpService;
     }
 
     public async Task<KeyValuePair<int, Data>?> GetById(int id)
     {
         //try get from local storage if not get from clusters
         var res = await _dataService.GetById(id);
-        if (res.Value.Value == null)
-        {
-            res = await _httpService.GetById(id, Settings.Server1);
-        }
 
-        if (res.Value.Value == null)
+        if (Settings.Leader)
         {
-            res = await _httpService.GetById(id, Settings.Server2);
+            if (res.Value.Value == null)
+            {
+                res = await _httpService.GetById(id, Settings.Server2);
+            }
         }
 
         return res;
@@ -37,28 +38,16 @@ public class DistributionService : IDistributionService
 
     public async Task<IDictionary<int, Data>?> GetAll()
     {
-        //call ge all from all servers and remove duplicates
+        //call get all from all servers and remove duplicates
 
         var resultDictionary = await _dataService.GetAll();
 
-        //try to get from server 1 if not get from server 2
-        var server1Data = await _httpService.GetAll(Settings.Server1);
-
-        if (resultDictionary != null && server1Data != null)
+        if (Settings.Leader)
         {
-            foreach (var data in server1Data)
-            {
-                if (!resultDictionary.ContainsKey(data.Key))
-                {
-                    resultDictionary.Add(data);
-                }
-            }
-        }
-        else if(resultDictionary != null)
-        {
+            //try to get from server 1 if not get from server 2
             var server2Data = await _httpService.GetAll(Settings.Server2);
 
-            if (server2Data != null)
+            if (resultDictionary != null && server2Data != null)
             {
                 foreach (var data in server2Data)
                 {
@@ -76,8 +65,10 @@ public class DistributionService : IDistributionService
     public async Task<Data> Update(int id, Data data)
     {
         //update try update all servers
-        var server1Data = await _httpService.Update(id, data, Settings.Server1);
-        var server2Data = await _httpService.Update(id, data, Settings.Server2);
+        if (Settings.Leader)
+        {
+            var server2Data = await _httpService.Update(id, data, Settings.Server2);
+        }
 
         return await _dataService.Update(id, data);
     }
@@ -86,16 +77,21 @@ public class DistributionService : IDistributionService
     {
         var results = new List<ResultSummary>();
 
-        var optimalServerNames = StorageHelper.GetOptimalServers();
+        var result = await _dataService.Save(data);
+        result.UpdateServerStatus();
 
-        if (optimalServerNames.Contains(Settings.ServerName))
+        results.Add(result);
+
+        //use tcp to save data to other servers
+        if (Settings.Leader)
         {
-            var result = await _dataService.Save(data);
-            result.UpdateServerStatus();
-
-            results.Add(result);
+            var server2Response = _tcpService.TcpSave(data, Settings.Server2TcpSavePort);
+            if (server2Response != null)
+            {
+                server2Response.UpdateServerStatus();
+                results.Add(server2Response);
+            }
         }
-
 
         return results;
     }
@@ -106,19 +102,15 @@ public class DistributionService : IDistributionService
         var result = await _dataService.Delete(id);
         result.UpdateServerStatus();
 
-        var server1Result = await _httpService.Delete(id, Settings.Server1);
-        var server2Result = await _httpService.Delete(id, Settings.Server2);
-        
-        if (server1Result != null)
+        if (Settings.Leader)
         {
-            server1Result.UpdateServerStatus();
-            results.Add(server1Result);
-        }
-        
-        if (server2Result != null)
-        {
-            server2Result.UpdateServerStatus();
-            results.Add(server2Result);
+            var server2Result = await _httpService.Delete(id, Settings.Server2);
+
+            if (server2Result != null)
+            {
+                server2Result.UpdateServerStatus();
+                results.Add(server2Result);
+            }
         }
 
         return results;
